% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{boxplot_ecc}
\alias{boxplot_ecc}
\title{Boxplot yearly ECC time-series (highcharts)}
\usage{
boxplot_ecc(
  data,
  prob = 0.95,
  title = "Historical Claim Costs across Years",
  subtitle = "As percent of limit, 2001-2019 average ECC with VaR line in red",
  ...
)
}
\arguments{
\item{data}{output of \link{wc_ecc}}

\item{prob}{which percentile threshold(s) to draw on charts}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Boxplot yearly ECC time-series (highcharts)
}
\examples{
data <- data.table(
  year = rep(2001:2020, each=5),
  ecc = sample(1:100/100, 5*20, replace=T)
  )
boxplot_ecc(data)

}
