% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bs_themed.R
\name{bs_themed}
\alias{bs_themed}
\title{Custom Bootstrap theme}
\usage{
bs_themed(
  base_font = "'national-web-book'",
  heading_font = "'DM Serif Text'",
  font_size = 15,
  palette = pal,
  elevation = 0,
  rounded = FALSE,
  gradients = FALSE
)
}
\arguments{
\item{base_font}{The default typeface.}

\item{heading_font}{The typeface to be used for heading elements.}

\item{font_size}{base font size in pixel}

\item{palette}{color palette (list of Bootstrap colors) (see \link{pal})}

\item{elevation}{default element shadow depth (see \href{https://getbootstrap.com/docs/4.1/utilities/shadows/}{BS4 shadows})}

\item{rounded}{enable rounded elements (see \href{https://getbootstrap.com/docs/4.0/utilities/borders/}{BS4 borders})}

\item{gradients}{enable background gradients (see \href{https://getbootstrap.com/docs/4.0/utilities/colors/}{BS4 colors})}
}
\value{
Modified Bootstrap 4 theme
}
\description{
A modified Bootstrap 4 theme based on IWMI branding guidelines.
}
\examples{
if(interactive()) bs_theme_preview(bs_themed())

}
