% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate.R
\name{calculate}
\alias{calculate}
\title{Calculated Measures}
\usage{
calculate(data)
}
\arguments{
\item{data}{data.table with columns (see \link{DATA})}
}
\value{

}
\description{
Series of built-in formulas to derive composite measures from WA+ model
output (see this package's vignettes for more details). These measures
are identified in the variable catalog with \code{sheet = "index"} (see \link{META}).
}
\examples{

}
