% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DATA}
\alias{DATA}
\title{Modeled WA+ data cube}
\format{
A data.table with columns:
\itemize{
\item iso3
\item sheet
\item class
\item subclass
\item variable
\item date_start
\item date_end
\item year
\item month
\item unit
\item value
}
}
\usage{
DATA
}
\description{
A normalized data.table of WA+ variables generated by \link{data_etl}.
}
\examples{
names(DATA)
DATA[1:5]

}
\seealso{
\link{ZOI}
}
\keyword{datasets}
