% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_etl.R
\name{data_etl}
\alias{data_etl}
\title{Main data ETL}
\usage{
data_etl(
  iso3 = NULL,
  root = getOption("wa.data"),
  file = file.path("./data-raw/rds/data.rds")
)
}
\arguments{
\item{iso3}{3-letter country code of the river basin to ingest. If null, all basin
ISO3 listed in \link{ISO3} will be processed. Can be used to manually test new result
sets.}

\item{root}{base directory (local or remote) to scan for input data files (default:
package option \code{getOption("wa.data")} or environment variable \code{WA_DATA} if set)}

\item{file}{a connection or the name of the file where the R object is saved to. The
default behavior is to save the transformed dataset to this package's \code{./data-raw}
folder, and to include it as a built-in package dataset (this is done
automatically on package build, see \link{DATA}).}
}
\value{
normalized data.table with variables: -
}
\description{
Main ETL to ingest WA+ modeled result sets found at location defined by
\code{getOption("wa.data")}. This location may be specified at run time or via
environment variable \code{WA_DATA}. Exact location and structure for output data files
are still to be defined.
}
\details{
A \href{../articles/02_data-etl.html}{vignette} is available to describe the current approach.
}
\examples{
data <- data_etl("mli")

}
