% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_themed.R
\name{hc_themed}
\alias{hc_themed}
\title{Custom HighCharts theme}
\usage{
hc_themed(
  hc,
  title = NULL,
  subtitle = NULL,
  label = NULL,
  x = NULL,
  y = NULL,
  base_font = "'national-web-book'",
  axes = TRUE,
  exporting = TRUE,
  credits = FALSE,
  ...
)
}
\arguments{
\item{hc}{a highchart object}

\item{title}{chart title}

\item{subtitle}{chart subtitle}

\item{label}{optional annotation}

\item{base_font}{default font for all chart elements}

\item{axes}{show axes (default: TRUE)}

\item{exporting}{include chart export menu (default: TRUE)}

\item{credits}{include credits (default: FALSE)}

\item{...}{
  Arguments passed on to \code{\link[highcharter:hc_theme]{highcharter::hc_theme}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
Custom HighCharts theme
}
