% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hc_themed.R
\name{hc_themed_vb}
\alias{hc_themed_vb}
\title{Custom Highcharts theme for sparklines}
\usage{
hc_themed_vb(hc, ...)
}
\arguments{
\item{hc}{a highchart object}

\item{...}{
  Arguments passed on to \code{\link[highcharter:hc_theme]{highcharter::hc_theme}}
  \describe{
    \item{\code{}}{}
  }}
}
\description{
This theme is used inside value boxes.
}
\examples{
data <- data.table(
  x = Sys.Date() + 0:9,
  y = rnorm(10)
)

hchart(data, "line", hcaes(x, y)) \%>\%
  hc_themed_vb() \%>\%
  hc_chart(backgroundColor=pal[["red"]])

}
