% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{hc_ts}
\alias{hc_ts}
\title{Plot dekadal index time series with trends (highcharts)}
\usage{
hc_ts(
  data,
  type = c("precipitation", "temperature"),
  yrange = NULL,
  title = "Long-Term Cumulative Rainfall by Dekad",
  subtitle = "Average across portfolio locations, 1990-2020 (mm)",
  ...
)
}
\arguments{
\item{data}{biophysical time-series (output of \link{wc_extract})}

\item{type}{precipitation or temperature variable}

\item{yrange}{y-axis min and max}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Plot dekadal index time series with trends (highcharts)
}
\examples{
dt <- readRDS(file.path(getOption("wc.shared"), "tmp", "2020-burn_lockton_bio.rds"))
plot_ts(dt[code=="bom_rain"])

}
