% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ISO3}
\alias{ISO3}
\title{River basins}
\format{
A named list, names are 3-letter ISO3 country codes.
}
\usage{
ISO3
}
\description{
A named list of basin metadata. Names are the basin's largest country ISO3 codes.
This list is read in from package configuration file at
\code{./data-raw/json/ISO3.json}. This file may be manually edited to add new basin
configurations.
}
\examples{
# Processed river basins
names(ISO3)

# Metadata terms currently in use for each basin
ISO3[[1]]

}
\seealso{
\link{ZOI}
}
\keyword{datasets}
