% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{l_admin}
\alias{l_admin}
\title{Format district picklist}
\usage{
l_admin(iso3 = names(ISO3))
}
\arguments{
\item{iso3}{Basin ISO3 code (see \link{ISO3})}
}
\value{
list of html tags
}
\description{
Format district picklist
}
\examples{
l_admin("ken")

}
