% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAYERS}
\alias{LAYERS}
\title{WMS and map tile providers}
\format{
A named list, names are provider codes.
}
\usage{
LAYERS
}
\description{
Named list of 3rd-party geospatial layers. By default layers are read in from a
package configuration file at \code{./data-raw/json/LAYERS.json}. This file may be
manually edited to add new layers to the dashboard. All spatial formats supported
by the \href{https://leafletjs.com/}{Leaflet} JS viewer are allowed (e.g. raster
tiles, WMS, GeoJSON, KML, etc.). Only \strong{vector tiles} are not supported currently.
}
\details{
The list is organized by data provider (e.g. Digital Earth Africa, ESRI World
Atlas, FAO WaPOR, Google Earth Engine, etc.).
}
\examples{
LAYERS[[2]]

}
\keyword{datasets}
