% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{lmap_init}
\alias{lmap_init}
\title{Initialize map (Mapbox GL)}
\usage{
lmap_init(
  iso3 = names(ISO3),
  mapbox_key = getOption("wa.mapbox"),
  maptiler_key = getOption("wa.maptiler")
)
}
\arguments{
\item{iso3}{3-letter country code to center the map on (see \link{ISO3})}

\item{mapbox_key}{Mapbox API key (default to \code{getOption("wa.mapbox")})}

\item{maptiler_key}{Maptiler API key (default to \code{getOption("wa.maptiler")})}
}
\value{
Mapbox GL map widget with default layers
}
\description{
Load basin boundaries with default configuration.
}
\examples{
# Default rendering
map_init()

}
