% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapbox.R
\name{lmap_update}
\alias{lmap_update}
\title{Update map (Mapbox GL)}
\usage{
lmap_update(m, iso3 = names(ISO3))
}
\arguments{
\item{m}{leaflet map}

\item{iso3}{3-letter country code to update the map (see \link{ISO3})}
}
\value{
Mapbox map widget
}
\description{
Add selected hydrological features across river basins.
}
