% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{map_init}
\alias{map_init}
\alias{map_update}
\alias{map_toggle}
\title{Initialize Leafet map and 3rd-party spatiotemporal layers}
\usage{
map_init(iso3 = names(ISO3), key = getOption("wa.maptiler"))

map_update(map, iso3 = names(ISO3))

map_toggle(map, layers = NULL)
}
\arguments{
\item{iso3}{3-letter country code to update the map (see \link{ISO3})}

\item{key}{Maptiler API key}

\item{map}{leaflet map}

\item{layers}{vector of layer names (see \link{LAYERS})}
}
\value{
leaflet map widget with default layers

leaflet map widget with default layers

updated leaflet map
}
\description{
Load basin boundaries with default configuration.

Add selected hydrological features across river basins.

Used to toggle 3rd-party contextual layers.
}
\examples{
# Default rendering
map_init()

}
