% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{META}
\alias{META}
\title{Variable catalog}
\format{
A data.table with columns:
\itemize{
\item sheet
\item id
\item class
\item subclass
\item variable
\item calculated
}
}
\usage{
META
}
\description{
A data.table of variable metadata (measure and dimension codes).
}
\examples{
names(META)
META[1:5]

}
\seealso{
\link{DATA}
}
\keyword{datasets}
