% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pal}
\alias{pal}
\title{Custom color palette}
\format{
A named character vector
}
\usage{
pal
}
\description{
A list of Bootstrap color codes. By default colors are read in from package
configuration file at \code{./data-raw/json/palette.gpl} (GIMP palette format).
}
\examples{
names(pal)
scales::show_col(pal, borders="#ffffff")

}
\keyword{datasets}
