% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_bullet}
\alias{plot_bullet}
\title{Plot bullet charts (highcharts)}
\usage{
plot_bullet(data, horizontal = TRUE, ...)
}
\arguments{
\item{data}{input data}

\item{horizontal}{boolean orientation}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot bullet charts (highcharts)
}
\examples{
data <- DATA[iso3=="mli" & sheet=="sheet1" & period=="year"]
plot_bullet(data)

}
