% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_clim}
\alias{plot_clim}
\title{Plot dekadal index time series (highcharts)}
\usage{
plot_clim(
  data,
  type = c("precipitation", "tmin", "tmax"),
  sos = switch(type, precipitation = 35, 2),
  yrange = NULL,
  title = "Climate Normals by Dekad",
  subtitle = "Average across portfolio locations, 1990-2020 (mm)",
  ...
)
}
\arguments{
\item{data}{biophysical time-series (output of \code{wc_extract()})}

\item{type}{index type \code{tp} (precipitation) or \code{temp} (temperature)}

\item{sos}{start of season threshold (default: 35 mm or 8°C Tmin for blooming)}

\item{yrange}{y-axis min and max}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Plot dekadal index time series (highcharts)
}
\examples{
dt <- readRDS(file.path(getOption("wc.shared"), "tmp", "2020-burn_lockton_bio.rds"))
plot_clim(dt[code=="bom_rain"])

}
