% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_ecc}
\alias{plot_ecc}
\title{Plot yearly ECC time-series (highcharts)}
\usage{
plot_ecc(
  data,
  title = "Historical Claim Costs across Perils",
  subtitle = "As percent of limit, 2001-2019 average ECC",
  ...
)
}
\arguments{
\item{data}{output of \link{wc_ecc}}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Plot yearly ECC time-series (highcharts)
}
\examples{
data <- data.table(
  year = rep(2001:2020, each=5),
  ecc = sample(1:100/100, 5*20, replace=T),
  ecc_dr = sample(1:100/100, 5*20, replace=T),
  ecc_xs = sample(1:100/100, 5*20, replace=T),
  ecc_tp = sample(1:100/100, 5*20, replace=T),
  weight = sample(1:100/100, 5*20, replace=T)
  )
plot_ecc(data)

}
