% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_ecc_density}
\alias{plot_ecc_density}
\title{Plot yearly ECC density (highcharts)}
\usage{
plot_ecc_density(
  data,
  prob = 0.95,
  title = "Probability Density of Claim Costs",
  subtitle = "As percent of limit, 2001-2019 average ECC with \%s VaR",
  ...
)
}
\arguments{
\item{data}{output of \link{wc_ecc}}

\item{prob}{which percentile threshold(s) to draw}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Plot yearly ECC density (highcharts)
}
