% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gauge.R
\name{plot_luc}
\alias{plot_luc}
\title{Plot land uses (highcharts)}
\usage{
plot_luc(iso3 = names(ISO3), ...)
}
\arguments{
\item{iso3}{basin ISO3 code}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot land uses (highcharts)
}
\examples{
plot_luc("ken")

}
