% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_profile}
\alias{plot_profile}
\title{Plot time-series intra-annual variability (highcharts)}
\usage{
plot_profile(
  data,
  unit = NA,
  yrange = NULL,
  polar = FALSE,
  unit.threshold = 10,
  ...
)
}
\arguments{
\item{data}{data.table to plot with columns \code{date} and \code{value} (will use \code{date_start}
if \code{date} is missing)}

\item{unit}{display unit}

\item{yrange}{user supplied y-axis range}

\item{polar}{use polar (circular) axis}

\item{unit.threshold}{maximum threshold to convert from km続 to MCM unit (default: 10 km続)}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\value{
highcharter object
}
\description{
Plot time-series intra-annual variability (highcharts)
}
\examples{
dt <- DATA[iso3=="ken" & id=="rainfall" & period=="month"]
plot_profile(dt, unit="km続")
plot_profile(dt, unit="km続", polar=TRUE)

}
