% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_radar.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot radar charts (highcharts)}
\usage{
plot_radar(iso3 = names(ISO3), ...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot radar charts (highcharts)
}
\examples{
plot_radar(iso3="ken")

}
