% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_rcp}
\alias{plot_rcp}
\title{Plot product recipes (highcharts)}
\usage{
plot_rcp(
  args,
  x = NULL,
  title = "Coverage Periods and Payout Factors | \%s",
  subtitle = "Range of payout factors by index type (\% of S.I.)",
  ...
)
}
\arguments{
\item{args}{pricing arguments}

\item{x}{(optional) which location ID (loc_id) to draw}

\item{title}{chart title}

\item{subtitle}{chart subtitle}
}
\description{
Plot product recipes (highcharts)
}
\examples{
pts <- data.table(loc_id=1, X=1, Y=1, day="2020-03-02")
args <- wc_args(pts, crop="maiz", pay_grm=.3)
plot_rcp(args)

}
