% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gauge.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Plot Sankey (highcharts)}
\usage{
plot_sankey(data, unit = NA, colors = pal, ...)
}
\arguments{
\item{data}{(optional) data with columns \code{from}, \code{to}, \code{weight} and \code{color}}

\item{unit}{display unit}

\item{colors}{ordered color palette}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot Sankey (highcharts)
}
\examples{
plot_sankey(dt)

}
