% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_tick}
\alias{plot_tick}
\title{Plot payout curves (highcharts)}
\usage{
plot_tick(data, title = NA, subtitle = NA, rowheight = 200, ...)
}
\arguments{
\item{data}{output of \link{wc_price} at 1 trigger location (single contract)}

\item{title}{chart title}

\item{subtitle}{chart subtitle}

\item{rowheight}{Height in px.}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}

\item{args}{pricing arguments}
}
\description{
Plot payout curves (highcharts)
}
\examples{
pts <- data.table(loc_id=1, X=30, Y=-17, day="2020-12-01")
args <- wc_args(pts, product="MAIZ-ZWE-05-CB01", code="rfe")
data <- wc_price(args=args)
plot_tick(data, args)

}
