% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot monthly basin time series (highcharts)}
\usage{
plot_timeline(
  data,
  yrange = NULL,
  title = NULL,
  subtitle = NULL,
  unit = "km³",
  ...
)
}
\arguments{
\item{data}{biophysical time-series}

\item{yrange}{y-axis min and max}

\item{title}{chart title}

\item{subtitle}{chart subtitle}

\item{sos}{start of season threshold (default: 35 mm or 8°C Tmin for blooming)}
}
\description{
Plot monthly basin time series (highcharts)
}
\examples{

}
