% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_ts.R
\name{plot_ts}
\alias{plot_ts}
\title{Plot variable time-series}
\usage{
plot_ts(dt, color = pal[1], unit = "km³")
}
\arguments{
\item{unit}{}
}
\value{

}
\description{
Plot variable time-series
}
\examples{
plot_ts(DATA[iso3=="mli" & sheet=="sheet1" & variable=="Rainfall"])

}
