% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_waterfall}
\alias{plot_waterfall}
\title{Plot waterfall charts}
\usage{
plot_waterfall(data, ...)
}
\arguments{
\item{data}{input data}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot waterfall charts
}
\examples{
vars = c("Rainfall", "Utilized", "Protected", "Main riverstem")
data = DATA[iso3=="mli" & sheet=="sheet1" & variable \%in\% vars
][, .(value=sum(value, na.rm=T)), by=.(year, variable)
][variable \%in\% vars[2:4], value := -value]
plot_waterfall(data)

}
