% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valueBoxSpark.R
\name{valueBoxSpark}
\alias{valueBoxSpark}
\title{Create a value box with a sparkline}
\usage{
valueBoxSpark(
  value,
  unit = "",
  selected = NA,
  title = "",
  subtitle = "",
  info = NULL,
  sparkobj = NULL,
  type = NA,
  icon = NULL,
  color = "light",
  width = 4,
  href = NULL
)
}
\arguments{
\item{value}{The value to display in the box. Usually a number or short text.}

\item{unit}{display unit}

\item{selected}{selected \code{x} value to highlight}

\item{subtitle}{Subtitle text.}

\item{info}{tooltip information text}

\item{sparkobj}{(optional) highcharts object}

\item{type}{(optional) type of spark chart (line, column, etc.) default to NA (no sparkline)
if not passed}

\item{icon}{An icon tag, created by \code{\link[shiny]{icon}}.}

\item{color}{The color of the item. This determines the item's background
  color.  Valid colors are defined as follows:
\itemize{

  \item \code{primary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#007bff")}.
  \item \code{secondary}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6c757d")}.
  \item \code{info}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#17a2b8")}.
  \item \code{success}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#28a745")}.
  \item \code{warning}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ffc107")}.
  \item \code{danger}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#dc3545")}.
  \item \code{gray-dark}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#343a40")}.
  \item \code{gray}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#adb5bd")}.
  \item \code{white}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#fff")}.
  \item \code{indigo}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#6610f2")}.
  \item \code{lightblue}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3c8dbc")}.
  \item \code{navy}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#001f3f")}.
  \item \code{purple}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#605ca8")}.
  \item \code{fuchsia}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#f012be")}.
  \item \code{pink}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#e83e8c")}.
  \item \code{maroon}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#d81b60")}.
  \item \code{orange}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#ff851b")}.
  \item \code{lime}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#01ff70")}.
  \item \code{teal}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#39cccc")}.
  \item \code{olive}: \Sexpr[results=rd, stage=render]{bs4Dash:::rd_color_tag("#3d9970")}.
}}

\item{width}{The width of the box, using the Bootstrap grid system. This is
used for row-based layouts. The overall width of a region is 12, so the
default width of 4 occupies 1/3 of that width. For column-based
layouts, use \code{NULL} for the width; the width is set by the column that
contains the box.}

\item{href}{An optional URL to link to in the footer. Should both `footer`
and this parameter be set, `footer` will take precedence.}

\item{data}{(optional) data.table to summarize with at least columns \code{x} and \code{y}}
}
\value{

}
\description{
Modified version of \code{\link[bs4Dash:valueBox]{bs4Dash::valueBox()}}
}
\examples{
data <- data.table(
  x = Sys.Date() + 0:9,
  y = rnorm(10)
)
valueBoxSpark(data, title="Title", subtitle="Subtitle", color="success", type="line")

data <- DATA[iso3=="ken" & id=="net_inflow" & period=="year", .(year, value)]
valueBoxSpark(data,
  title="Title", subtitle="Subtitle",
  color="success", type="bar", unit="km3")

valueBoxSpark(33,
  title="Title", subtitle="Subtitle",
  color="success", unit="km3", footer="Footer")

}
