% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WADashboard.R
\docType{package}
\name{WADashboard-package}
\alias{WADashboard}
\alias{WADashboard-package}
\title{WADashboard: IWMI Water Accounting+ Visualization Dashboard}
\description{
This package includes all dependencies to run WA+ visualization dashboard either locally inside RStudio or containerized.
}
\details{
R package with Shiny-based Water Accounting Plus dashboard application

This package exposes these options:
\itemize{
\item \code{wa.data}: user-defined path to download remote spatio-temporal raster files
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mbacou/WADashboard/}
  \item \url{https://iwmi.cgiar.org/}
  \item Report bugs at \url{https://github.com/mbacou/WADashboard/issues/}
}

}
\author{
\strong{Maintainer}: Melanie Bacou \email{mel@mbacou.com} (\href{https://orcid.org/0000-0003-1810-6818}{ORCID})

Authors:
\itemize{
  \item Lisa-Maria Rebelo \email{L.REBELO@CGIAR.ORG} (\href{https://orcid.org/0000-0002-8785-7810}{ORCID}) [contributor]
  \item William Rex \email{W.Rex@cgiar.org} (\href{https://orcid.org/0000-0003-4979-5257}{ORCID}) [contributor]
  \item Naga Velpuri \email{N.Velpuri@cgiar.org} (\href{https://orcid.org/0000-0002-6370-1926}{ORCID}) [contributor]
}

}
\keyword{internal}
