% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ZOI}
\alias{ZOI}
\title{Basin boundaries}
\format{
A named list, names are 3-letter ISO3 country codes.
}
\usage{
ZOI
}
\description{
Named list of simplified basin boundaries and water stream networks as \code{sf}
objects, read in from GeoJSON file locations defined in \link{ISO3}.
}
\examples{
par(mfrow=c(1, length(ZOI)))
for(i in ZOI) {
 plot(st_geometry(i[["admin"]]), col=pal["light"])
 plot(st_geometry(i[["water"]]), col=pal["blue"], add=T)
}

}
\seealso{
\link{ISO3}
}
\keyword{datasets}
