% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highcharts.R
\name{plot_radar}
\alias{plot_radar}
\title{Plot radar charts}
\usage{
plot_radar(data, ...)
}
\arguments{
\item{data}{input data}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{subtitle}}{chart subtitle}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot radar charts
}
\examples{
data = DATA[iso3=="mli" & sheet=="sheet1"]
plot_radar(data)

}
