% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gauge.R
\name{plot_wheel}
\alias{plot_wheel}
\title{Plot dependency wheel (highcharts)}
\usage{
plot_wheel(data, unit = NA, colors = pal, icons = NA, rot = 180, subtitle, ...)
}
\arguments{
\item{data}{(optional) data with columns \code{from}, \code{to}, \code{weight}, \code{color} and \code{icon}
(e.g. "tint")}

\item{unit}{display unit}

\item{colors}{ordered color palette}

\item{rot}{start angle}

\item{subtitle}{chart subtitle}

\item{...}{
  Arguments passed on to \code{\link[=hc_themed]{hc_themed}}
  \describe{
    \item{\code{hc}}{a highchart object}
    \item{\code{title}}{chart title}
    \item{\code{label}}{optional annotation}
    \item{\code{base_font}}{default font for all chart elements}
    \item{\code{axes}}{show axes (default: TRUE)}
    \item{\code{exporting}}{include chart export menu (default: TRUE)}
    \item{\code{credits}}{include credits (default: FALSE)}
  }}
}
\description{
Plot dependency wheel (highcharts)
}
\examples{
plot_wheel(dt)

}
